% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rock_palettes.R
\name{rock_palette}
\alias{rock_palette}
\title{Color Palettes based on classic rock album covers}
\usage{
rock_palette(name, n, type = c("discrete", "continuous"))
}
\arguments{
\item{name}{Name of palette. Select one:
\code{alice}, \code{californication}, \code{coltrane}, \code{deelite},
\code{electric}, \code{facelift}, \code{faithnomore}, \code{harvey},
\code{heep}, \code{hellawaits}, \code{husker}, \code{janelle},
\code{melloncollie},\code{miles}, \code{muse}, \code{nodoubt},
\code{peacesells},\code{secondlaw}, \code{siamesedream}, \code{swift},
\code{zeppelin}}

\item{n}{Number of colors desired.

Some palettes contain 5 colors. Those beginning with 'rock' have 4}

\item{type}{Either continuous or discrete.}
}
\value{
A vector of colors.
}
\description{
R package that contains color palettes based on colours on classic rock album covers.
}
\details{
See also: https://github.com/johnmackintosh/metallicaRt for metallica palettes
}
\examples{
rock_palette("swift")

}
